fixed: reset timeslotUrl when time range is selected manually
