+++ js/app/components/delivery-form/TimeSlotPicker.js
@@ -1,80 +1,113 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useMemo, useCallback } from 'react'
 import { DatePicker, Select, Radio } from 'antd'
 import moment from 'moment'
-import { useFormikContext } from 'formik'
 import { useTranslation } from 'react-i18next'
 
 import './TimeSlotPicker.scss'
 import Spinner from '../core/Spinner'
+import {
+  useDeliveryFormFormikContext
+} from './hooks/useDeliveryFormFormikContext'
+import { useGetStoreQuery } from '../../api/slice'
 
 const baseURL = location.protocol + '//' + location.host
 
-export default ({ index, timeSlotLabels }) => {
+const extractDateAndRangeFromTimeSlot = (timeSlotChoice) => {
+  let [first, second] = timeSlotChoice.split('/')
+  first = moment(first)
+  second = moment(second)
+  const date = moment(first)
+  const hour = `${first.format('HH:mm')}-${second.format('HH:mm')}`
+  return { date, hour }
+}
 
-  const httpClient = new window._auth.httpClient()
+const formatSlots = (timeSlotChoices) => {
+  const formattedSlots = {}
+  timeSlotChoices.forEach(choice => {
+    const { date, hour } = extractDateAndRangeFromTimeSlot(choice.value)
+    const formattedDate = date.format('YYYY-MM-DD')
+    if (formattedSlots[formattedDate]) {
+      formattedSlots[formattedDate].push(hour)
+    } else {
+      formattedSlots[formattedDate] = [hour]
+    }
+  })
+
+  return formattedSlots
+}
+
+export default ({ storeId, index, timeSlotLabels }) => {
+
+  const httpClient = useMemo(() => {
+    return new window._auth.httpClient()
+  }, [])
+
+  const { data: store } = useGetStoreQuery(storeId)
+
+  const timeSlotIds = store?.timeSlots
+  const defaultTimeSlotId = store?.timeSlot
 
   const { t } = useTranslation()
 
-  const { setFieldValue, values } = useFormikContext()
+  const { isModifyOrderMode, taskValues, setFieldValue } = useDeliveryFormFormikContext({
+    taskIndex: index,
+  })
 
   const [formattedTimeslots, setFormattedTimeslots] = useState({})
   const [isLoadingChoices, setIsLoadingChoices] = useState(false)
 
-  const extractDateAndRangeFromTimeSlot = (timeSlotChoice) => {
-    let [first, second] = timeSlotChoice.split('/')
-    first = moment(first)
-    second = moment(second)
-    const date = moment(first)
-    const hour = `${first.format('HH:mm')}-${second.format('HH:mm')}`
-    return { date, hour }
-  }
-
-  const formatSlots = (timeSlotChoices) => {
-    const formattedSlots = {}
-    timeSlotChoices.forEach(choice => {
-      const { date, hour } = extractDateAndRangeFromTimeSlot(choice.value)
-      const formattedDate = date.format('YYYY-MM-DD')
-      if (formattedSlots[formattedDate]) {
-        formattedSlots[formattedDate].push(hour)
-      } else {
-        formattedSlots[formattedDate] = [hour]
-      }
-    })
-
-    return formattedSlots
-  }
-
-  const setTimeSlotFromDateAndRange = (date, hourRange) => {
+  const setTimeSlotFromDateAndRange = useCallback((date, hourRange) => {
     const [first, second] = hourRange.split('-')
     const startTime = date.clone().hours(first.split(':')[0]).minutes(first.split(':')[1])
     const endTime = date.clone().hours(second.split(':')[0]).minutes(second.split(':')[1])
     const timeSlot = `${startTime.utc().format('YYYY-MM-DDTHH:mm:00')}Z/${endTime.utc().format('YYYY-MM-DDTHH:mm:00')}Z`
     setFieldValue(`tasks[${index}].timeSlot`, timeSlot)
-  }
-
-  const getTimeSlotChoices = async timeSlotUrl => {
-    setIsLoadingChoices(true)
-
-    const url = `${baseURL}${timeSlotUrl}/choices`
-    const { response } = await httpClient.get(url)
-
-    if (response['choices'].length === 0) {
-      setFieldValue(`tasks[${index}].timeSlot`, 'No choice')
-    } else {
-      const formattedSlots = formatSlots(response['choices'])
-      setFormattedTimeslots(formattedSlots)
-
-      const availableDates = Object.keys(formattedSlots)
-      const firstDate = moment(availableDates[0])
-      setTimeSlotFromDateAndRange(firstDate, formattedSlots[availableDates[0]][0])
-    }
-
-    setIsLoadingChoices(false)
-  }
+  }, [index, setFieldValue])
 
   useEffect(() => {
-    getTimeSlotChoices(values.tasks[index].timeSlotUrl)
-  }, [values.tasks[index].timeSlotUrl])
+    const getTimeSlotChoices = async timeSlotUrl => {
+      setIsLoadingChoices(true)
+
+      const url = `${baseURL}${timeSlotUrl}/choices`
+      const { response } = await httpClient.get(url)
+
+      if (response['choices'].length === 0) {
+        setFieldValue(`tasks[${index}].timeSlot`, null)
+      } else {
+        const formattedSlots = formatSlots(response['choices'])
+        setFormattedTimeslots(formattedSlots)
+
+        const availableDates = Object.keys(formattedSlots)
+        const firstDate = moment(availableDates[0])
+        setTimeSlotFromDateAndRange(firstDate, formattedSlots[availableDates[0]][0])
+      }
+
+      setIsLoadingChoices(false)
+    }
+
+    if (!taskValues.timeSlotUrl) {
+      return
+    }
+
+    getTimeSlotChoices(taskValues.timeSlotUrl)
+  }, [taskValues.timeSlotUrl, index, setFieldValue, httpClient, setTimeSlotFromDateAndRange])
+
+  useEffect(() => {
+    if (isModifyOrderMode) {
+      return
+    }
+
+    if (taskValues.timeSlotUrl) {
+      return
+    }
+
+    if (timeSlotIds.length === 0) {
+      return
+    }
+
+    setFieldValue(`tasks[${index}].timeSlotUrl`, defaultTimeSlotId ?? timeSlotIds[0])
+
+  }, [timeSlotIds, defaultTimeSlotId, index, taskValues.timeSlotUrl, isModifyOrderMode, setFieldValue])
 
   const handleTimeSlotLabelChange = e => {
     setFieldValue(`tasks[${index}].timeSlot`, null)
@@ -82,18 +115,18 @@ export default ({ index, timeSlotLabels }) => {
   }
 
   const handleDateChange = newDate => {
-    const { hour } = extractDateAndRangeFromTimeSlot(values.tasks[index].timeSlot)
+    const { hour } = extractDateAndRangeFromTimeSlot(taskValues.timeSlot)
     setTimeSlotFromDateAndRange(newDate, hour)
   }
 
   const handleHourChange = hourRange => {
-    const { date } = extractDateAndRangeFromTimeSlot(values.tasks[index].timeSlot)
+    const { date } = extractDateAndRangeFromTimeSlot(taskValues.timeSlot)
     setTimeSlotFromDateAndRange(date, hourRange)
   }
 
   const inputLabel = () => <div className="mb-2 font-weight-bold title-slot">{t('ADMIN_DASHBOARD_FILTERS_TAB_TIMERANGE')}</div>
 
-  if (isLoadingChoices || !values.tasks[index].timeSlot) {
+  if (isLoadingChoices || !taskValues.timeSlot) {
     return (
       <>
         {inputLabel()}
@@ -107,8 +140,8 @@ export default ({ index, timeSlotLabels }) => {
     return !availableDates.some(date => date.isSame(current, 'day'))
   }
 
-  const selectedDate = extractDateAndRangeFromTimeSlot(values.tasks[index].timeSlot).date
-  const selectedHour = extractDateAndRangeFromTimeSlot(values.tasks[index].timeSlot).hour
+  const selectedDate = extractDateAndRangeFromTimeSlot(taskValues.timeSlot).date
+  const selectedHour = extractDateAndRangeFromTimeSlot(taskValues.timeSlot).hour
 
   const hourOptions = formattedTimeslots[selectedDate.format('YYYY-MM-DD')] || []
 
@@ -119,7 +152,7 @@ export default ({ index, timeSlotLabels }) => {
       {timeSlotLabels.length > 1 ?
         <Radio.Group
           className="timeslot__container mb-2"
-          value={values.tasks[index].timeSlotUrl}
+          value={taskValues.timeSlotUrl}
         >
           {timeSlotLabels.map(label => (
             <Radio.Button
@@ -135,7 +168,7 @@ export default ({ index, timeSlotLabels }) => {
         : null
       }
 
-      { hourOptions.length > 0 ? 
+      { hourOptions.length > 0 ?
         <div style={{ display: 'flex', marginTop: '0.5em' }}>
           <DatePicker
             format="LL"
