+++ js/app/components/delivery-form/redux/store.js
@@ -1,0 +1,15 @@
+import { configureStore } from '@reduxjs/toolkit'
+import { accountSlice } from '../../../entities/account/reduxSlice'
+import { apiSlice } from '../../../api/slice'
+
+export function createStoreFromPreloadedState(preloadedState) {
+  return configureStore({
+    reducer: {
+      [accountSlice.name]: accountSlice.reducer,
+      [apiSlice.reducerPath]: apiSlice.reducer,
+    },
+    preloadedState,
+    middleware: getDefaultMiddleware =>
+      getDefaultMiddleware().concat(apiSlice.middleware),
+  })
+}
